refactor: display registration success snackbar on home page
